moving downloadContracts into download-contracts service
