+++ src/services/index.mjs
@@ -1,5 +1,5 @@
-import post_verified_addresses from "./handle-verified-address.mjs";
-import download from "./download.mjs";
+import postVerifiedAddresses from "./handle-verified-address.mjs";
+import downloadContracts from "./download-contract.mjs";
 import writeMetrics from "./write-metrics.mjs";
 
-export { post_verified_addresses, download, writeMetrics };
+export { postVerifiedAddresses, downloadContracts, writeMetrics };
