+++ src/utils/index.mjs
@@ -1,3 +1,4 @@
 import delayLoop from "./delay-loop.mjs";
+import urlDirFn from "./url-dir-fn.mjs";
 
-export { delayLoop };
+export { delayLoop, urlDirFn };
